implementation of quicksort
ended up using these as a resources for understanding

https://en.wikipedia.org/wiki/Quicksort
http://blog.benoitvallon.com/sorting-algorithms-in-javascript/the-quicksort-algorithm/

for a while my termination control in my partition was off by one, I wasn't
scanning the entire array. I was going up to endPosition -1 in the
partition function in my for loop. once that was corrected, partition
worked as expected.